Organized code better
